Splitted booking mailing service to 3
