--- Api/Infrastructure/Options/PaymentNotificationOptions.cs
@@ -1,7 +1,0 @@
-namespace HappyTravel.Edo.Api.Infrastructure.Options
-{
-    public class PaymentNotificationOptions
-    {
-        public string ReceiptTemplateId { get; set; }
-    }
-}
