+++ HappyTravel.Edo.UnitTests/Tests/Services/Accommodations/Bookings/BookingsProcessingServiceTests/Charging/ExecutingCharging.cs
@@ -5,9 +5,9 @@ using CSharpFunctionalExtensions;
 using HappyTravel.Edo.Api.Models.Users;
 using HappyTravel.Edo.Api.Services.Accommodations.Bookings;
 using HappyTravel.Edo.Api.Services.Accommodations.Bookings.BatchProcessing;
+using HappyTravel.Edo.Api.Services.Accommodations.Bookings.Mailing;
 using HappyTravel.Edo.Api.Services.Accommodations.Bookings.Management;
 using HappyTravel.Edo.Api.Services.Accommodations.Bookings.Payments;
-using HappyTravel.Edo.Api.Services.Mailing;
 using HappyTravel.Edo.Common.Enums;
 using HappyTravel.Edo.Data.Bookings;
 using HappyTravel.Edo.Data.Management;
@@ -111,7 +111,8 @@ namespace HappyTravel.Edo.UnitTests.Tests.Services.Accommodations.Bookings.Booki
             var service = new BookingsProcessingService(bookingPaymentService,
                 Mock.Of<IBookingCreditCardPaymentService>(),
                 bookingManagementService,
-                Mock.Of<IBookingMailingService>(),
+                Mock.Of<IBookingNotificationService>(),
+                Mock.Of<IBookingReportsService>(),
                 context.Object);
             return service;
         }
